Transfer code from sidebar to layout
